3.1.1 and 3.1.2 - Created IntArrayStack class and changed DefaultCountingOutRhymer
