+++ src/main/java/edu/kis/vh/nursery/FirstInFirstOutRhymer.java
@@ -5,7 +5,7 @@ public class FirstInFirstOutRhymer extends DefaultCountingOutRhymer {
 	private final DefaultCountingOutRhymer temp = new DefaultCountingOutRhymer();
 
 	@Override
-	protected int countOut() {
+	public int countOut() {
 		while (!callCheck())
 		    temp.countIn(super.countOut());
